Make the dita source generation threaded
This dramatically speeds up the generation of the source as we load each webpage concurrently.
The SpecificationParser is a Callable task which is sent to the threaded executor service which is limited to the usual `Runtime.runtime.availableProcessors().intdiv(2)`.

Also include a cpl of method tidy ups and logging to help track where the generation is up to.