Rename the Dita Map class to be generated as a class called DitaMap
This solves a major issue for users using the DSL where there is a collision between the java.util.Map and the dita map.
The collision requires users to either use an `import ... as ...` statement for the Dita map,
or use the FQDN for the java Map.
Both approaches are just annoying and unnecessary when we control the code generation